Lecture 6 material

+++ P&P2_FPS/Assets/Scripts/EnemyAI.cs
@@ -18,7 +18,7 @@ public class EnemyAI : MonoBehaviour, IDamage
     private Animator m_animator = null;
 
     [SerializeField]
-    private int m_health = 100;
+    private float m_health = 100;
     [SerializeField]
     private int m_faceTargetSpeed = 20;
     [SerializeField]
@@ -136,7 +136,7 @@ public class EnemyAI : MonoBehaviour, IDamage
         return false;
     }
     
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         m_health -= amount;
         m_navMeshAgent.SetDestination(GameManager.Instance.m_player.transform.position);
